requirements.txt
discord.py==2.0.0
mysql-connector-python==8.0.28
pytz==2021.3
python-dotenv==0.19.2
requests==2.26.0
